Some improvements after code review.
