+++ stream-compositions/api/integrations-api/legal-entity-integration-api/pom.xml
@@ -51,7 +51,7 @@
                             <goal>validate</goal>
                         </goals>
                         <configuration>
-                            <input>src/main/resources/service-api-v1.yaml</input>
+                            <input>src/main/resources/integration-api-v1.yaml</input>
                             <failOnWarning>true</failOnWarning>
                         </configuration>
                     </execution>
