+++ stream-compositions/services/legal-entity-composition-service/src/main/java/com/backbase/stream/compositions/legalentity/LegalentityCompositionApplication.java
@@ -2,15 +2,14 @@ package com.backbase.stream.compositions.legalentity;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.boot.web.servlet.support.SpringBootServletInitializer;
 import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
 import org.springframework.context.annotation.ComponentScan;
 
 @EnableDiscoveryClient
 @SpringBootApplication
 @ComponentScan("com.backbase.stream")
-public class Application extends SpringBootServletInitializer {
+public class LegalentityCompositionApplication {
     public static void main(String[] args) {
-        SpringApplication.run(Application.class, args);
+        SpringApplication.run(LegalentityCompositionApplication.class, args);
     }
 }
