+++ stream-compositions/services/legal-entity-composition-service/src/main/resources/application.yml
@@ -38,7 +38,7 @@ backbase:
       legal-entity:
         enable-completed-events: true
         enable-failed-events: true
-        legal-entity-integraion-url: http://legal-entity-ingestion-integration:8080
+        legal-entity-integration-url: http://legal-entity-ingestion-integration:8080
 
 bootstrap:
   # This is just na example. Replace it with data corresponding to your bank
